Add config file for CI
Add the config file necessary to get this to work with CircleCI 2. I
know ahead of time that this build is going to fail tests (because there
are a block of tests that I never actually finished), and it may fail
during the checkout stage because git isn't installed on the image.
We'll see.