Add Racket metadata file
Add the metadata file (`info.rkt`) for Racket. The particular advantage
that this gives me is to specify dependencies. Currently I am only
relying on the deps that are always installed with Racket (at the
moment), but I will very shortly be adding dependencies on some other
libraries in order to write the parser more neatly.