Redefine tokens
Previously I was passing things around in form of a string, which is a
very sub-par way of passing around information.  This changes tokens
from being strings (eg `":"`) to symbols (`'duplicate`), which has the
associated advantage of being their own types. Has an additional
advantage with push tokens, in that now I can deal with stripping parens
a little more nicely.